Working version, need cleanup: approot property +example
