+++ src/BlazorWorker.WorkerCore/JSInvokeException.cs
@@ -18,20 +18,4 @@ namespace BlazorWorker.WorkerCore
         {
         }
     }
-
-    [Serializable]
-    public class WorkerException : Exception
-    {
-        public WorkerException(string message) : base(message)
-        {
-        }
-
-        public WorkerException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
-
-        protected WorkerException(SerializationInfo info, StreamingContext context) : base(info, context)
-        {
-        }
-    }
 }
